% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_colors.R
\name{color_map}
\alias{color_map}
\title{A mapping function to specific colors.}
\usage{
color_map(name)
}
\arguments{
\item{name}{a character (vector) corresponding to a
name with a mappable color}
}
\value{
a character (vector) of hex codes
}
\description{
This function maps a common feature/cell
type (e.g. "HSC") and returns the canonical
hex code associtated with it.
}
\details{
color_map (no 's') is the function; color_maps
is a named character vector.
}
\examples{
# See possible values
names(color_maps)

# Get hex codes
color_map("HSC")
color_map(c("HSC", "LMPP", "HSC"))

}
\keyword{colors}
