% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_colors.R
\docType{data}
\name{ejc_color_maps}
\alias{ejc_color_maps}
\title{List of maps between annotations and specified colors for scHeme/UKBB
project}
\format{An object of class \code{character} of length 19.}
\usage{
ejc_color_maps
}
\description{
Use ejc_color_map to map cell/feature to color hex.
}
\keyword{datasets}
