% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHelp.R
\name{g_legend}
\alias{g_legend}
\title{Extract a legend from ggplot}
\usage{
g_legend(a.gplot)
}
\arguments{
\item{a.gplot}{ggplot plot object with a color / fill legend}
}
\value{
A new ggplot object with only the legend
}
\description{
Grab the legend from a ggplot object and use it
for saving in its own file
}
\examples{
df <- data.frame(x = rnorm(1000), y = 0)
p1 <- ggplot(shuf(df), aes(x=x, y=y, colour=x)) + geom_point()
legendP <- g_legend(p1)

}
\keyword{guide}
\keyword{legend}
