% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHelp.R
\name{get_density}
\alias{get_density}
\title{Compute the density of points in a grid}
\usage{
get_density(x, y, n = 200)
}
\arguments{
\item{x}{X coordinates}

\item{y}{Y coordinates}

\item{n}{Number of bins to divide grid}
}
\value{
A vector of densities for plotting
}
\description{
Credit to Kamil Slowikowski
See post: http://slowkow.com/notes/ggplot2-color-by-density/
}
\examples{
dat <- data.frame(
  x = c(
    rnorm(1e4, mean = 0, sd = 0.1),
    rnorm(1e3, mean = 0, sd = 0.1)
  ),
  y = c(
    rnorm(1e4, mean = 0, sd = 0.1),
    rnorm(1e3, mean = 0.1, sd = 0.2)
  )
)
dat$density <- get_density(dat$x, dat$y)
ggplot2::ggplot(dat) + geom_point(aes(x, y, color = density))
}
\keyword{density}
\keyword{point}
