% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_colors.R
\docType{data}
\name{jdb_color_maps}
\alias{jdb_color_maps}
\title{List of maps between annotations and specified colors.}
\format{An object of class \code{character} of length 18.}
\usage{
jdb_color_maps
}
\description{
Use \code{\link{jdb_color_map}} to map cell/feature to color hex.
}
\keyword{datasets}
