% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{jdb_palette}
\alias{jdb_palette}
\title{A palette generator}
\usage{
jdb_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are: lots.}

\item{n}{Number of colors desired.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes for the Buenrostrolab.
}
\examples{
jdb_palette("Cavalcanti")
jdb_palette("Cavalcanti", 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- jdb_palette(21, name = "Zissou", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}

