% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{jdb_palettes}
\alias{jdb_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 105.}
\usage{
jdb_palettes
}
\description{
Use \code{\link{jdb_palette}} to construct palettes of desired length.
}
\keyword{datasets}
