% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHelp.R
\name{L_border}
\alias{L_border}
\title{Axes along the bottom and left}
\usage{
L_border()
}
\value{
A ggplot theme object
}
\description{
Modify ggplot theme such that
only the bottom X axis and left Y axis are shown making an "L"
}
\examples{
df <- data.frame(x = rnorm(1000), y = 0)
p1 <- ggplot(shuf(df), aes(x=x, y=y, colour=x)) + geom_point()
p1

}
\keyword{guide}
\keyword{legend}
