% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHelp.R
\name{numberToColorVec}
\alias{numberToColorVec}
\title{Maps numeric vector to color palette}
\usage{
numberToColorVec(numberVec, palette, range = NULL)
}
\arguments{
\item{numberVec}{Numeric vector}

\item{palette}{color palette to which to be mapped. Any
valid palette in "BuenColors" should work.}

\item{range}{The c(min, max) numeric values to
make the color at the extremes.}
}
\description{
Modified from Sushi::maptocolors
}
\examples{
i <- rnorm(1000)
df <- data.frame(x = i, y = 0, color = numberToColorVec(i^2, "brewer_heat"))
ggplot(shuf(df), aes(x=x, y=y, colour= color)) + geom_point() +
 scale_color_manual(values = levels(df$color))

}
