% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHelp.R
\name{pretty_plot}
\alias{pretty_plot}
\title{Get rid of gridlines, etc.}
\usage{
pretty_plot()
}
\value{
A vector of densities for plotting
}
\description{
Idea is to make it look more like Matlab
}
\examples{
df <- data.frame(x = rnorm(1000), y = 0)
ggplot(df, aes(x=x, y=y, colour=x)) + geom_point()

}
\keyword{density}
\keyword{point}
