% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHelp.R
\name{shuf}
\alias{shuf}
\title{Randomly shuffle the plotting order}
\usage{
shuf(df)
}
\value{
An object with the rows shuffled
}
\description{
Idea is to make it look more like
}
\examples{
df <- data.frame(x = rnorm(1000), y = 0)
ggplot(shuf(df), aes(x=x, y=y, colour=x)) + geom_point()

}
\keyword{density}
\keyword{point}
