% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_colors.R
\name{tf1_color_map}
\alias{tf1_color_map}
\title{A mapping function to specific colors for the
TF1 clone paper}
\usage{
tf1_color_map(name)
}
\arguments{
\item{name}{a character (vector) corresponding to a
name with a mappable color}
}
\value{
a character (vector) of hex codes
}
\description{
This function maps a common feature/cell
type (e.g. "B9") and returns the canonical
hex code associtated with it.
}
\details{
color_map (no 's') is the function; color_maps
is a named character vector.
}
\examples{
# See possible values
names(tf1_color_maps)

# Get hex codes
tf1_color_map("B9")
tf1_color_map(c("C9", "D6", "G10"))

}
\keyword{colors}
