% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_colors.R
\docType{data}
\name{tf1_color_maps}
\alias{tf1_color_maps}
\title{List of maps between annotations and specified colors
for the TF1 phylogeny analyses.}
\format{An object of class \code{character} of length 17.}
\usage{
tf1_color_maps
}
\description{
Use tf1_color_map to map cell/feature to color hex.
}
\keyword{datasets}
